+++ src/util/serializing/ObjectOutputStream.cpp
@@ -50,19 +50,19 @@ void ObjectOutputStream::writeString(const char* str) { writeString(std::string(
 
 void ObjectOutputStream::writeString(const std::string& s) {
     this->encoder->addStr("_s");
-    int len = static_cast<int>(s.length());
-    this->encoder->addData(&len, sizeof(int));
+    size_t len = s.length();
+    this->encoder->addData(&len, sizeof(size_t));
     this->encoder->addData(s.c_str(), len);
 }
 
-void ObjectOutputStream::writeData(const void* data, size_t len, int width) {
+void ObjectOutputStream::writeData(const void* data, size_t len, size_t width) {
     this->encoder->addStr("_b");
-    this->encoder->addData(&len, sizeof(int));
+    this->encoder->addData(&len, sizeof(size_t));
 
     // size of one element
-    this->encoder->addData(&width, sizeof(int));
+    this->encoder->addData(&width, sizeof(size_t));
     if (data != nullptr) {
-        this->encoder->addData(data, static_cast<int>(len) * width);
+        this->encoder->addData(data, len * width);
     }
 }
 
@@ -70,7 +70,7 @@ void ObjectOutputStream::writeImage(const std::string_view& imgData) {
     this->encoder->addStr("_m");
     size_t len = imgData.length();
     this->encoder->addData(&len, sizeof(size_t));
-    this->encoder->addData(imgData.data(), static_cast<int>(len));
+    this->encoder->addData(imgData.data(), len);
 }
 
 auto ObjectOutputStream::getStr() -> GString* { return this->encoder->getData(); }
